+++ spec/models/dorsale/customer_vault/activity_type_spec.rb
@@ -1,3 +1,5 @@
+require "rails_helper"
+
 RSpec.describe ::Dorsale::CustomerVault::ActivityType, type: :model do
   it { is_expected.to validate_presence_of :name }